requirements.txt
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
python-dotenv==1.0.0